generated_app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
snowflake-connector-python==3.6.0
pydantic==2.5.3
python-multipart==0.0.6
